xtensa: add support for the configurable Xtensa architecture
The Xtensa processor architecture is a configurable, extensible,
and synthesizable 32-bit RISC processor core. Processor and SOC vendors
can select from various processor options and even create customized
instructions in addition to a base ISA to tailor the processor for
a particular application.

Because of the configurability, the build process requires one additional
step for gcc, binutils, and gdb to update the default configuration.
These configurations are packed into an 'overlay' tar image, and are
simply untarred on top of the default configuration during the build.

This patch was originally authored by Chris Zankel. This version fixes a few
issues raised by Yann E. MORIN, and tested on latest tip.

Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>